+++ returns/pointfree.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from returns._generated.pointfree.bind import _bind as bind  # noqa: F401
+from returns._generated.pointfree.rescue import _rescue as rescue  # noqa: F401