+++ tests/test_maybe/test_maybe_equality.py
@@ -3,13 +3,7 @@
 import pytest
 
 from returns.maybe import Nothing, Some, _Nothing
-from returns.primitives.container import (
-    Bindable,
-    Fixable,
-    Mappable,
-    Rescueable,
-    Unwrapable,
-)
+from returns.primitives.container import Bindable, Mappable, Unwrapable
 from returns.primitives.exceptions import ImmutableStateError
 
 
@@ -20,8 +14,6 @@ from returns.primitives.exceptions import ImmutableStateError
 @pytest.mark.parametrize('protocol', [
     Bindable,
     Mappable,
-    Fixable,
-    Rescueable,
     Unwrapable,
 ])
 def test_protocols(container, protocol):
@@ -52,9 +44,6 @@ def test_is_compare():
     some_container = Some(1)
 
     assert Nothing.bind(lambda state: state) is Nothing
-    assert some_container.rescue(  # type: ignore
-        lambda _: Some('fix'),
-    ) is some_container
     assert some_container is not Some(1)